requirements.txt
asgiref==3.10.0
attrs==25.4.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
Faker==38.0.0
faker-commerce==1.0.4
gunicorn==23.0.0
h11==0.16.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
model-bakery==1.20.5
packaging==25.0
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
uvicorn==0.38.0
